Push modal onto presentedViewController
When trying to use this library on top of a modal it silently fails
because it's trying to present on top of the root view which already has
a modal on top of it.

This changes reactRoot to a getter which will return the correct
UIViewController to push onto, either the root or the currently
presented modal.