Handle nil tokenizedPayPalAccount
When the user closes the paypal window via the `Done` button or `Cancel
Sandbox Purchase` both `error` and `tokenizedPayPalAccount` are nil
which throws an exception.

To handle this, we now explicitly check for truthiness of
`tokenizedPayPalAccount` before the success branch and check if `error`
is not nil before trying to return the error description.

By default we return nil in cases where there is no description.